==> question_answering/02_chat_only.ipynb <==
# 02. Chat Only

Attempt to use only the chat data to answer questions

> As we can see, the concept is not bad, but we cannot send the entire chat history to OpenAI, because it has a limit of 4096 tokens. So we need to find a way to summarize the chat history or pass just relevant text.

import os
import pandas as pd
import openai
from dotenv import load_dotenv

from IPython.core.display import Markdown

## Configure Environment

reference: [Using .env Files for Environment Variables in Python Applications](https://dev.to/jakewitcher/using-env-files-for-environment-variables-in-python-applications-55a1)

load_dotenv()

OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
OPENAI_CHAT_MODEL = os.getenv('OPENAI_CHAT_MODEL', 'gpt-3.5-turbo')

openai.api_key = OPENAI_API_KEY

## Load Data

df_source = pd.read_csv('data/output/transcripts.csv')

print(df_source.shape)
with pd.option_context('display.max_columns', None):
    display(df_source.head(3))

## Functions

def create_session_history(data: pd.DataFrame) -> dict:
    session_history = []

    for index, row in data.iterrows():
        session_history.append({
            'role': 'user',
            'content': row['transcript']
        })

    return session_history

#create_session_history(df_source)

def ask_question(question: str, session_history: list) -> str:
    messages = session_history.copy()

    # add the user message
    messages.append({
        'role': 'user',
        'content': question
    })

    response = openai.ChatCompletion.create(
        model=OPENAI_CHAT_MODEL,
        messages=messages
    )

    return response['choices'][0]['message']['content']

# test the function
question = 'Reinforcement learning might be described as learning a Q function. If we are using Q-learning to learn how to play video games, state TWO aspects that make Q-learning difficult.'
#ask_question(question, create_session_history(df_source))

## Summarize

ask = f'Summarize the following: {df_source.iloc[101].transcript}'
summary = ask_question(ask, [])

print(summary)

transcript = df_source.iloc[43].transcript
word_limit = 1000

#ask = f'Keep as much detail as possible in {word_limit} words: {transcript}'
#ask = f'could you give me a summary of this text in around {word_limit} words? {transcript}'
#ask = f'Produce a list of bullet points of the knowledge in this text: {transcript}'
ask = f'provide each concept explained as a bullet list: {transcript}'

summary = ask_question(ask, [])

# print the transcript and the summary length
print('Transcript length:', len(transcript))
print('Summary length:', len(summary))

# print the transcript
print('--- Transcript ---')
display(Markdown(transcript))

# print the summary
print('--- Summary ---')
display(Markdown(summary))

## Search Keywords

# suggest search keywords
ask = f'suggest search keywords for the following as a comma seperated list: {question}'
response = ask_question(ask, [])

keywords = response.replace('\n', '').split(',')
print(keywords)
