==> question_answering/04_pinecone_query.ipynb <==
# 04. Pinecone Query

Query the index created in notebook 03.

import os
import openai
import pinecone

from tqdm.auto import tqdm
from dotenv import load_dotenv
from IPython.core.display import Markdown

## Configure Environment

load_dotenv()

OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
OPENAI_CHAT_MODEL = os.getenv('OPENAI_CHAT_MODEL', 'gpt-3.5-turbo')
OPENAI_EMBED_MODEL = os.getenv('OPENAI_EMBED_MODEL', 'text-embedding-ada-002')

PINECONE_API_KEY = os.getenv('PINECONE_API_KEY')
PINECONE_INDEX = os.getenv('PINECONE_INDEX', 'openai-dsm100-2022-oct-transcriptions')
PINECONE_ENV = os.getenv('PINECONE_ENV', 'us-east1-gcp')

# configure openai
openai.api_key = OPENAI_API_KEY

# configure pinecone
pinecone.init(
    api_key=PINECONE_API_KEY,
    environment=PINECONE_ENV
)

## Define Sample Questions

question01 = """
a. AI has a long history with ups and downs, over-optimistic expectations, and the
industrial period. At present, AI is in the era of:

Select ALL statements that apply.

[3 marks]
i. Big data.
ii. Deep learning.
iii. Robotics.
iv. General AI.
"""

question02 = """
f. The more background knowledge an intelligent agent can use – the more intelligent
behaviour it can exhibit. Background knowledge can include:
[3 marks]

Select ALL statements that apply. 

(i). Problem-specific knowledge
(ii). Declarative knowledge
(iii). Procedural knowledge
(iv). A dataset
"""

## Functions

def get_pinecone_index(index_name: str) -> pinecone.index.Index:
    # connect to index
    index = pinecone.Index(index_name)
    
    # view index stats
    print(index.describe_index_stats())  

    return index  

# test the function
index = get_pinecone_index(PINECONE_INDEX)

def get_text_embeddings(text: str) -> list:
    result = openai.Embedding.create(
            input=[text], engine=OPENAI_EMBED_MODEL)

    return result['data'][0]['embedding']

# test the function
#get_text_embeddings(question01)

def query_pinecone_index(index: pinecone.index.Index,
                         query: str,
                         k: int = 5,
                         include_metadata=True) -> list:
    # get embeddings for query
    query_embedding = get_text_embeddings(query)

    # query index
    return index.query(
        vector=query_embedding,
        top_k=k,
        include_metadata=include_metadata
    )

# test the function
#query_pinecone_index(index, question02)

def create_prompt(index: pinecone.index.Index,
                  query: str, 
                  prompt_limit: int = 3750) -> str:
    prompt = ""
    
    # query pinecone for matching transcripts
    pinecone_results = query_pinecone_index(index, query)

    # get relevant contexts
    contexts = [
        x['metadata']['text'] for x in pinecone_results['matches']
    ]

    # build our prompt with the retrieved contexts included
    prompt_start = (
        "Answer the question based on the context below.\n\n"+
        "Context:\n"
    )
    
    prompt_end = (
        f"\n\nQuestion: {query}\nAnswer:"
    )

    # append contexts until hitting limit
    for i in range(1, len(contexts)):
        if len("\n\n---\n\n".join(contexts[:i])) >= prompt_limit:
            prompt = (
                prompt_start +
                "\n\n---\n\n".join(contexts[:i-1]) +
                prompt_end
            )
            break
        elif i == len(contexts)-1:
            prompt = (
                prompt_start +
                "\n\n---\n\n".join(contexts) +
                prompt_end
            )

    return prompt, pinecone_results

# test the function
prompt, context = create_prompt(index, question02)
display(Markdown(prompt))



def ask_question(index: pinecone.index.Index,
                  query: str, 
                  prompt_limit: int = 3750):

    # create prompt
    prompt, context = create_prompt(index, query, prompt_limit)

    response = openai.ChatCompletion.create(
        model=OPENAI_CHAT_MODEL,
        messages=[
            { 'role': 'user', 'content': prompt }
        ]
    )

    return response['choices'][0]['message']['content'], context

# test the function
answer, _ = ask_question(index, question02)
display(Markdown('Question:' + question02))
display(Markdown('Answer: ' + answer))

answer, _ = ask_question(index, question01)
display(Markdown('Question:' + question01))
display(Markdown('Answer: ' + answer))

## Question Definitions

def display_answer(index: pinecone.index.Index,
                   question: str):
   answer, context = ask_question(index, question)

   output = []

   output.append('### Question')
   output.append(question)

   output.append('### Answer')
   output.append(answer)

   output.append('### Context')
   context_list = set(x['metadata']['topic'] + '::' + x['metadata']['title'] for x in context['matches'])
   for item in context_list:
      output.append('- ' + item)

   display(Markdown('\n'.join(output)))

# test the function
display_answer(index, question01)


questions = {}

question = """
AI has a long history with ups and downs, over-optimistic expectations, and the
industrial period. At present, AI is in the era of:
[3 marks]

Select ALL statements that apply.
i. Big data.
ii. Deep learning.
iii. Robotics.
iv. General AI.
"""

questions['A_1_a'] = question
display_answer(index, question)

question = """
b. It is important for AI systems to be explainable and interpretable because it:
Select ALL statements that apply.
[3 marks]

i. Gives users confidence in the system.
ii. Is a legal requirement.
iii. Helps in safeguarding against bias.
iv. Enables assessing of system’s robustness and vulnerability. 
"""

questions['A_1_b'] = question
display_answer(index, question)

question = """
c. Heuristic functions can be generated based on:
Select ALL statements that apply.
[3 marks]

i. Simplified (or relaxed) problems.
ii. Forecasted blind search results.
iii. Previous experience.
iv. Features of the search space.
"""

questions['A_1_c'] = question
display_answer(index, question)

question = """
c. Heuristic functions can be generated based on:
Select ALL statements that apply.
[3 marks]

i. Simplified (or relaxed) problems.
ii. Forecasted blind search results.
iii. Previous experience.
iv. Features of the search space.
"""

questions['A_1_c'] = question
display_answer(index, question)

question = """
e. A genetic algorithm:
Select ALL statements that apply.
[3 marks]

i. Is a global search algorithm
ii. Is a local search algorithm
iii. Is inspired by Charles Darwin’s theory of natural evolution.
iv. Includes three evolutionary processes: selection, gene crossover, and
mutation.
"""

questions['A_1_e'] = question
display_answer(index, question)

question = """
f. The more background knowledge an intelligent agent can use – the more intelligent
behaviour it can exhibit. Background knowledge can include:
Select ALL statements that apply.
[3 marks]

(i). Problem-specific knowledge
(ii). Declarative knowledge
(iii). Procedural knowledge
(iv). A dataset
"""

questions['A_1_f'] = question
display_answer(index, question)

question = """
g. An intelligent agent needs to handle uncertainty only when the environment is both:
partially observable and nondeterministic
[3 marks]

Is this statement is True/False?
"""

questions['A_1_g'] = question
display_answer(index, question)

question = """
The principle of maximum expected utility (MEU) says that a rational agent should
choose:
Select ONE most appropriate answer.
[3 marks]

(i). The action that maximises the agent’s current utility.
(ii). The action that maximises the agent’s expected utility.
(iii). The utility function that maximises the expected agent’s value.
(iv). The utility function that maximises the current agent’s value.
"""

questions['A_1_h'] = question
display_answer(index, question)

question = """
i. The Bayesian approach to the source and status of probability numbers is:
Select ONE most appropriate answer.
[3 marks]

(i). Subjective.
(ii). Objective.
(iii). Frequentist.
(iv). Statistical.
"""

questions['A_1_i'] = question
display_answer(index, question)

question = """
j. A typical cycle of scientific discovery includes the following steps:
Select ALL statements that apply.
[3 marks]

i. Conducting surveys.
ii. Hypothesis generation.
iii. Deduction – a prediction from the hypothesis about future effects or
experimental results
iv. Hypothesis testing, to verify or disprove predictions.
"""

questions['A_1_j'] = question
display_answer(index, question)

question = """
a) What type of problems are Markov Decision Processes (MDPs) suitable for
modelling?
"""

questions['B_2_a'] = question
display_answer(index, question)

question = """
b) Choose a video game that is significantly different from the Breakout game
seen in the module.
[2 marks]

• Name and briefly describe the game you have chosen.
• What are the control inputs?
• How can you gain points in the game?
• How might an artificial intelligence system perceive the state of the
game? Please do not use the example of screen pixels - think about
the underlying state of the game. 
"""

questions['B_2_b'] = question
display_answer(index, question)

question = """
c) Name and describe THREE elements of a Markov Decision Process. For
each element, say what it represents from your chosen video game.
[3 marks]
"""

questions['B_2_c'] = question
display_answer(index, question)

question = """
d) Do you think a Markov Decision Process is a suitable way to model playing
your chosen game? Justify your answer.
[2 marks]
"""

questions['B_2_d'] = question
display_answer(index, question)

question = """
e) Reinforcement learning might be described as learning a Q function. If we
are using Q-learning to learn how to play video games, state TWO aspects
that make Q-learning difficult.
[2 marks]
"""

questions['B_2_e'] = question
display_answer(index, question)

question = """
Nowadays AI systems are capable of making autonomous scientific discoveries. In
this module, we considered several AI systems capable of making automated
scientific discoveries, for example, the mobile robotic chemist, the automatic
statistician, the robot scientists Adam and Eve. The research in automated scientific
discovery is flourishing, and the target is to develop systems capable of making
significant scientific discoveries.

a) What are the advantages of automated discovery systems?
[3 marks]
"""

questions['B_3_a'] = question
display_answer(index, question)

question = """
Nowadays AI systems are capable of making autonomous scientific discoveries. In
this module, we considered several AI systems capable of making automated
scientific discoveries, for example, the mobile robotic chemist, the automatic
statistician, the robot scientists Adam and Eve. The research in automated scientific
discovery is flourishing, and the target is to develop systems capable of making
significant scientific discoveries.

b) What are the current limitations of automated discovery systems? 
[3 marks]
"""

questions['B_3_b'] = question
display_answer(index, question)

question = """
Nowadays AI systems are capable of making autonomous scientific discoveries. In
this module, we considered several AI systems capable of making automated
scientific discoveries, for example, the mobile robotic chemist, the automatic
statistician, the robot scientists Adam and Eve. The research in automated scientific
discovery is flourishing, and the target is to develop systems capable of making
significant scientific discoveries.

c) Describe the functionality of one of the key components of the robot scientist
system, e.g. a hypotheses generator, a reasoning engine, an experiments
planner, or data analysis component. What is required for the functioning of
the selected component? What are its input and output? How does it interact
with other components of the system? Outline the workings of the component.
[4 marks]
"""

questions['B_3_c'] = question
display_answer(index, question)

question = """
a) What is a genetic algorithm?
[1 mark]
"""

questions['B_4_a'] = question
display_answer(index, question)

question = """
b) Draw a diagram which illustrates the runtime of a genetic algorithm. Show the
major components of the software and when they are used.
[5 marks]
"""

questions['B_4_b'] = question
display_answer(index, question)

question = """
c) Genetic algorithms employ a design process which is similar to the human
design process. Provide TWO arguments against this statement and TWO
arguments in support of this statement, using details of the design of genetic
algorithms.
[4 marks]
"""

questions['B_4_c'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). A
description of an AI agent should include specification of its tasks, environment,
background knowledge (if any), actions, performance measures.

Propose an AI agent. For example, a birthday agent can monitor when birthdays of
your friends, relatives and colleagues are, advise you what birthday presents to buy
depending on preferences and availability, organise their purchase and delivery.
A description of your AI agent must include the following:

• Aim: what is the purpose for the development of your AI agent? What tasks
will it be it able to work on?
[10 marks]
"""

questions['C_5_a'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages).

• Motivation: why the proposed agent is important to develop? Who will
benefit from the development of the proposed AI agent?
[5 marks]
"""

questions['C_5_b'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). 

Environment: describe the environment where your agent will be acting. Is it
deterministic, not deterministic, observable, partially observable, etc.?
Explain, why do you think so? 
[10 marks]
"""

questions['C_5_c'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). 

• Functionality: briefly describe the workings of your agent. Will be it a standalone system? Can it be integrated with other AI agents or systems?
[5 marks]
"""

questions['C_5_d'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). 

Data and knowledge: Will your agent use dataset(s) and knowledge? If so,
describe those data, knowledge and how they can be obtained. If not,
explain why using data and knowledge is not required.
[5 marks]
"""

questions['C_5_e'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). 

• Evaluation: Explain what performance measures can be used to evaluate
your agent.
[5 marks]
"""

questions['C_5_g'] = question
display_answer(index, question)

question = """
In this module, we have adopted an agent-based approach for considered AI systems.
Your task is to propose an AI agent and provide its description (~1-2 pages). A
description of an AI agent should include specification of its tasks, environment,
background knowledge (if any), actions, performance measures.

Propose an extension or a modification of one of the considered agents or AI
systems. For example, you can modify the vacuum cleaner example by proposing an
intelligent vacuum cleaner that can decide what to clean, how and when. It can be
integrated with an intelligent shopping agent for buying cleaning materials.
Integration with a calendar agent can help with scheduling of cleaning, for example
before the owner is back from holidays. 

A description of your modification or extension must include the following:
• A brief description of the original agent or system: its tasks, functionality,
environment, its limitations.
[5 marks]

• Aim of the modified or extended agent: what is the purpose for the
development of your agent? What additional tasks will it be it able to work on?
[5 marks]

• Environment: describe the environment where your agent will be acting. Is it
deterministic, not deterministic, observable, partially observable, etc.? Is it
different from the original environment? Explain, why do you think so? 
[10 marks]

• Functionality: briefly describe the workings of your modified/extended agent.
What is the difference with the original agent/system?
[10 marks]

• Data and knowledge: Will your modified/ extended agent use additional
dataset(s) and knowledge? If so, describe those data, knowledge and how
they can be obtained. If not, explain why using data and knowledge is not
required.
[5 marks]

• Evaluation: Explain what performance measures can be used to evaluate your
modified/ extended agent.
[5 marks]
"""

questions['C_6'] = question
display_answer(index, question)
